traefik/pyproject.toml
[project]
name = "traefik"
version = "0.1.0"
description = ""
readme = "readme.md"
requires-python = ">=3.14"
dependencies = [
    "cytoolz>=1.0.1",
    "more-itertools>=10.7.0",
    "rich>=14.0.0",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project.scripts]
gen = "traefik.gen:gen"
clean = "traefik.gen:clean"
genclean = "traefik.gen:genclean"

[tool.mypy]
packages = ["traefik"]
exclude = ["deprecated"]
ignore_missing_imports = true

[tool.ty.environment]
root = ["./src"]
